shopinvader_api_sale: move search schema to shopinvader_schema_sale
