+++ shopinvader_api_sale/__init__.py
@@ -1,2 +1,1 @@
 from . import routers
-from . import schemas
