--- shopinvader_api_sale/schemas.py
@@ -1,16 +1,0 @@
-# Copyright 2023 Akretion (https://www.akretion.com).
-# @author Sébastien BEAU <sebastien.beau@akretion.com>
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
-
-
-from odoo.addons.extendable_fastapi import StrictExtendableBaseModel
-
-
-class SaleSearch(StrictExtendableBaseModel):
-    name: str | None = None
-
-    def to_odoo_domain(self):
-        if self.name:
-            return [("name", "ilike", self.name)]
-        else:
-            return []
